Use token array instead of stack
Return token array ("program") from pn_parse and rename pn_stack_eval to
pn_eval and STACK_SIZE to MAX_TOKENS to match the new usage.

A stack is not needed for representing the parsed expression since it
was only ever pushed to before the first pop. An array will suffice, but
it has to be reversed to put the tokens in the right order for
evaluation.

Not using a stack is possibly more efficient and simplifies the code.
The stack implementation can now be focused around the value stack.