Remove size check from stack_push and make void
The v_stack can never be full as it's of size MAX_TOKENS and there can
not be more values than that. Therefore the check to see if the stack is
full is not needed, and the return value from stack_push was ignored
anyway. The return type was changed to void.

This will simplify code and make calls to stack_push slightly more
efficient.

The function stack_is_full was added for good measure, if the stack is
to be used in another way later, where the max number of elements is
unknown.