Fix major realloc bug in stack_push
The size of the new array was computed incorrectly as stack->size * 2
bytes, instead of stack->size * 2 ints. This lead to major issues if the
input had more than the initial amount of values allocated in the stack.

pcalc is now tested to work with over 4k values in one expression. I
tested this using a simple shell script.